Move Auth library httpclient implementation to its own file
