+++ source/com.microsoft.tfs.client.common/src/com/microsoft/tfs/client/common/credentials/CredentialsHelper.java
@@ -3,19 +3,15 @@
 
 package com.microsoft.tfs.client.common.credentials;
 
-import java.io.IOException;
 import java.net.URI;
 import java.text.MessageFormat;
-import java.util.HashMap;
 import java.util.Locale;
-import java.util.Map;
 import java.util.TimeZone;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
 import com.microsoft.alm.auth.Authenticator;
-import com.microsoft.alm.auth.HttpClientFactory;
 import com.microsoft.alm.auth.PromptBehavior;
 import com.microsoft.alm.auth.oauth.DeviceFlowResponse;
 import com.microsoft.alm.auth.oauth.Global;
@@ -23,9 +19,6 @@ import com.microsoft.alm.auth.oauth.OAuth2Authenticator;
 import com.microsoft.alm.auth.pat.VstsPatAuthenticator;
 import com.microsoft.alm.client.TeeClientHandler;
 import com.microsoft.alm.helpers.Action;
-import com.microsoft.alm.helpers.HttpClient;
-import com.microsoft.alm.helpers.HttpResponse;
-import com.microsoft.alm.helpers.StringContent;
 import com.microsoft.alm.secret.Token;
 import com.microsoft.alm.secret.TokenPair;
 import com.microsoft.alm.secret.TokenType;
@@ -41,17 +34,9 @@ import com.microsoft.tfs.core.TFSTeamProjectCollection;
 import com.microsoft.tfs.core.credentials.CachedCredentials;
 import com.microsoft.tfs.core.credentials.CredentialsManager;
 import com.microsoft.tfs.core.httpclient.Credentials;
-import com.microsoft.tfs.core.httpclient.Header;
-import com.microsoft.tfs.core.httpclient.HttpMethodBase;
-import com.microsoft.tfs.core.httpclient.HttpStatus;
 import com.microsoft.tfs.core.httpclient.JwtCredentials;
 import com.microsoft.tfs.core.httpclient.PreemptiveUsernamePasswordCredentials;
 import com.microsoft.tfs.core.httpclient.UsernamePasswordCredentials.PatCredentials;
-import com.microsoft.tfs.core.httpclient.methods.EntityEnclosingMethod;
-import com.microsoft.tfs.core.httpclient.methods.GetMethod;
-import com.microsoft.tfs.core.httpclient.methods.HeadMethod;
-import com.microsoft.tfs.core.httpclient.methods.PostMethod;
-import com.microsoft.tfs.core.httpclient.methods.StringRequestEntity;
 import com.microsoft.tfs.core.util.URIUtils;
 import com.microsoft.tfs.jni.helpers.LocalHost;
 import com.microsoft.tfs.util.StringUtil;
@@ -306,135 +291,10 @@ public abstract class CredentialsHelper {
         }
     }
 
-    private static class AuthLibHttpClientImpl implements HttpClient {
-
-        final private com.microsoft.tfs.core.httpclient.HttpClient apacheClient;
-        final private Map<String, String> headers;
-
-        public AuthLibHttpClientImpl() {
-            final TFSTeamProjectCollection rootConnection = new TFSTeamProjectCollection(
-                URIUtils.VSTS_ROOT_URL,
-                null,
-                new CommonClientConnectionAdvisor(Locale.getDefault(), TimeZone.getDefault()));
-
-            apacheClient = rootConnection.getHTTPClient();
-            headers = new HashMap<String, String>();
-        }
-
-        /**
-         * {@inheritDoc}
-         */
+    private static class AuthLibHttpClientFactory extends com.microsoft.alm.auth.HttpClientFactory {
         @Override
-        public Map<String, String> getHeaders() {
-            return headers;
-        }
-
-        private String getResponseText(final HttpMethodBase request) throws IOException {
-            final int statusCode = apacheClient.executeMethod(request);
-            final String responseText = request.getResponseBodyAsString();
-
-            if (HttpStatus.isSuccessFamily(statusCode)) {
-                return responseText;
-            }
-
-            throw new IOException(HttpStatus.getStatusText(statusCode) + ":" + responseText); //$NON-NLS-1$
-        }
-
-        /**
-         * Timeout is ignored
-         */
-        @Override
-        public String getGetResponseText(final URI uri, final int timeout) throws IOException {
-            return getGetResponseText(uri);
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override
-        public String getGetResponseText(final URI uri) throws IOException {
-            final HttpMethodBase request = new GetMethod(uri.toString());
-            addHeader(request);
-
-            return getResponseText(request);
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override
-        public String getHeaderField(final URI uri, final String header) throws IOException {
-            final HttpMethodBase request = new HeadMethod(uri.toString());
-            addHeader(request);
-
-            apacheClient.executeMethod(request);
-
-            final Header[] headers = request.getResponseHeaders(header);
-            if (headers.length > 0) {
-                final Header responseHeader = headers[0];
-                return responseHeader.getValue();
-            }
-
-            return null;
-        }
-
-        private EntityEnclosingMethod createPostMethod(final URI uri, final StringContent content) {
-            final EntityEnclosingMethod request = new PostMethod(uri.toString());
-
-            request.setRequestEntity(new StringRequestEntity(content.getContent()));
-            getHeaders().putAll(content.Headers);
-            addHeader(request);
-
-            return request;
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override
-        public HttpResponse getPostResponse(final URI uri, final StringContent content) throws IOException {
-            final EntityEnclosingMethod request = createPostMethod(uri, content);
-
-            final int statusCode = apacheClient.executeMethod(request);
-
-            String responseOut = null;
-            String responseError = null;
-
-            if (HttpStatus.isSuccessFamily(statusCode)) {
-                responseOut = request.getResponseBodyAsString();
-            } else {
-                responseError = request.getResponseBodyAsString();
-            }
-
-            HttpResponse httpResponse = new HttpResponse();
-            httpResponse.status = statusCode;
-            httpResponse.errorText = responseError;
-            httpResponse.responseText = responseOut;
-
-            return httpResponse;
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override
-        public String getPostResponseText(URI uri, StringContent content) throws IOException {
-            final EntityEnclosingMethod request = createPostMethod(uri, content);
-
-            return getResponseText(request);
-        }
-
-        private void addHeader(final HttpMethodBase request) {
-            for (final Map.Entry<String, String> entry : headers.entrySet()) {
-                request.setRequestHeader(entry.getKey(), entry.getValue());
-            }
-        }
-    }
-
-    private static class AuthLibHttpClientFactory extends HttpClientFactory {
-        @Override
-        public HttpClient createHttpClient() {
-            return new AuthLibHttpClientImpl();
+        public com.microsoft.alm.helpers.HttpClient createHttpClient() {
+            return new AuthenticationHttpClientImpl();
         }
     }
 }
